requirements.txt
Django==1.8.7
requests==2.7.0
django-bootstrap3==6.1.0
django-registration-redux==1.2
jsonfield==1.0.0
lxml
vecnet.openmalaria
vecnet.simulation